refactor: Extract core/hooks.rs inline tests to external modular structure
- Extract 16 test functions (~1,320 lines) from core/hooks.rs
- Create modular test structure in src/tests/core_hooks/
- Organize tests by functionality: actions, config, context, execution, patterns
- Add test_utils module for shared test helpers
- Make necessary functions public for external test access
- Fix type mismatches and API signatures in extracted tests
- Remove duplicate module loading to resolve clippy warnings
- Maintain 100% test coverage with 565 tests passing

File size reduction: core/hooks.rs reduced from 2,244 to 925 lines (58% reduction)
