+++ src/tests/core_hooks/mod.rs
@@ -1,4 +1,5 @@
 pub mod actions;
+pub mod config;
 pub mod context;
 pub mod execution;
 pub mod patterns;
